Ajuste de grafico/indicador

+++ src/main/java/br/com/ottimizza/dashboard/models/graficos/GraficoDashboard.java
@@ -1,6 +1,7 @@
 package br.com.ottimizza.dashboard.models.graficos;
 
 import java.io.Serializable;
+import java.math.BigInteger;
 import lombok.AllArgsConstructor;
 import lombok.NoArgsConstructor;
 import lombok.Getter;
@@ -10,7 +11,7 @@ import lombok.Setter;
 @NoArgsConstructor @AllArgsConstructor
 public class GraficoDashboard implements Serializable {
     
-    private Long id;
+    private BigInteger id;
     private String nome;
     private Long abertoNoPrazo;
     private Long abertoAtrasado;
